feat(VTU): detect and use connectivity
After that, we will compute one tree per CC.
